# specs/003-import-from-storygraph-goodreads/contracts/import-api.openapi.yaml
openapi: 3.0.3
info:
  title: Tome Import API
  description: |
    API for importing reading history from TheStoryGraph and Goodreads CSV exports.
    
    **Features:**
    - Upload and parse CSV files from supported providers
    - Intelligent book matching using ISBN, fuzzy title/author matching
    - Preview matched/unmatched records before committing
    - Create reading sessions and sync ratings to Calibre
    - Track unmatched records for manual review
    
    **Workflow:**
    1. POST /api/import/upload - Upload CSV, get preview
    2. GET /api/import/:importId/preview - Review detailed matches
    3. POST /api/import/:importId/execute - Confirm and execute import
    4. GET /api/import/:importId/unmatched - Review unmatched records
  version: 1.0.0
  contact:
    name: Tome Support
    url: https://github.com/yourusername/tome
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://tome.example.com
    description: Production server
...
paths:
...
  /api/import/{importId}/unmatched:
    get:
      tags:
        - import
      summary: Get unmatched records for an import
      description: |
        Retrieve all unmatched records for a specific import. Useful for:
        - Exporting unmatched records to CSV
        - Manual matching via UI
        - Understanding matching failures
        - Debugging import issues
        
        **Common Unmatched Reasons:**
        - no_isbn: Book has no ISBN and title/author didn't match
        - no_title_match: Title similarity below threshold (70%)
        - ambiguous: Multiple potential matches found
        - not_in_library: Book not in Calibre library
      operationId: getUnmatchedRecords
      parameters:
        - name: importId
          in: path
          required: true
          description: UUID of the import
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        - name: format
          in: query
          required: false
          description: Response format (json or csv)
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Unmatched records retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnmatchedResponse'
              example:
                success: true
                importId: "550e8400-e29b-41d4-a716-446655440000"
                totalUnmatched: 14
                unmatched:
                  - id: 42
                    title: "Obscure Indie Novel"
                    authors: ["Unknown Author"]
                    isbn: null
                    rating: 4
                    completedDate: "2024-01-10T00:00:00Z"
                    status: "read"
                    review: "Great story but hard to find"
                    reason: "no_isbn"
                    matchAttempted: true
                    createdAt: "2024-12-01T10:35:00Z"
                  - id: 43
                    title: "The Art of Computer Programming Vol 1"
                    authors: ["Donald Knuth"]
                    isbn: "9780201896831"
                    rating: 5
                    completedDate: "2024-02-15T00:00:00Z"
                    status: "read"
                    reason: "not_in_library"
                    matchAttempted: true
                    createdAt: "2024-12-01T10:35:00Z"
            text/csv:
              schema:
                type: string
                format: binary
              example: |
                title,authors,isbn,rating,completedDate,status,reason
                "Obscure Indie Novel","Unknown Author",,"4","2024-01-10","read","no_isbn"
                "The Art of Computer Programming Vol 1","Donald Knuth","9780201896831","5","2024-02-15","read","not_in_library"
        '404':
          description: Import not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    UnmatchedResponse:
      type: object
      required:
        - success
        - importId
        - totalUnmatched
        - unmatched
      properties:
        success:
          type: boolean
          example: true
        importId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        totalUnmatched:
          type: integer
          minimum: 0
          example: 14
        unmatched:
          type: array
          items:
            $ref: '#/components/schemas/UnmatchedRecord'
    UnmatchedRecord:
      type: object
      required:
        - id
        - title
        - authors
        - reason
        - matchAttempted
        - createdAt
      properties:
        id:
          type: integer
          description: Database ID of unmatched record
          example: 42
        title:
          type: string
          example: "Obscure Indie Novel"
        authors:
          type: array
          items:
            type: string
          example: ["Unknown Author"]
        isbn:
          type: string
          nullable: true
          example: null
        rating:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
          example: 4
        completedDate:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-10T00:00:00Z"
        status:
          type: string
          example: "read"
        review:
          type: string
          nullable: true
          example: "Great story but hard to find"
        reason:
          type: string
          enum: [no_isbn, no_title_match, ambiguous, not_in_library]
          example: "no_isbn"
        matchAttempted:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-12-01T10:35:00Z"
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - code
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
          example: "Invalid CSV format"
        details:
          type: string
          description: Additional error details
          example: "Missing required columns: Date Read, My Rating"
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_CSV_FORMAT
            - FILE_TOO_LARGE
            - EMPTY_CSV
            - PROVIDER_MISMATCH
            - IMPORT_NOT_FOUND
            - IMPORT_ALREADY_EXECUTED
            - INVALID_MATCH_IDS
            - EXECUTION_FAILED
          example: "INVALID_CSV_FORMAT"
        fieldErrors:
          type: object
          additionalProperties:
            type: string
          description: Field-specific validation errors
          example:
            file: "File must be a CSV"
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session token (automatically sent by browser)
security:
  - sessionAuth: []